requirement.txt
appdirs==1.4.4
astroid==2.4.2
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
pathspec==0.8.1
pip==20.3.1
pkce==1.0
pylint==2.6.0
pymongo==3.11.2
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.0
setuptools==49.2.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
